Crate problem074

Problem 74 (Digit factorial chains)

Solution detail

For this problem, iterate up through the numbers 1, 2, 3, ..., 1,000,000. For each one, repeat the following process:

Create an empty hash set to keep track of terms in the chain so far

Repeatedly calculate the next term in the chain until:

If the next tern is in the hash set, then we have found a repeat, so can stop

If the next term is smaller than the starting value, we already know the chain length for this value, so can terminate early

Store the chain length for this starting value so that it can be used in the future as a shortcut for later values

Simply perform this algorithm until all chain lengths have been calculated, and keep track of how many chains of length 60 there are.

Functions

answer
Solve the problem, returning the answer as a String

factorial_digit_sum

find_chain_length

main

solve
Find how many numbers below the given limit form a chain of the given number of terms before repeating themselves.